+++ CHANGELOG.md
@@ -56,6 +56,8 @@ All notable changes to this project will be documented in this file.
 - Fixed sort agents by status in `GET/agents` API request. ([#810](https://github.com/wazuh/wazuh/pull/810))
 - Added exception when no agents are selected to restart. ([#870](https://github.com/wazuh/wazuh/pull/870))
 - Prevent files from remaining open in the cluster. ([#874](https://github.com/wazuh/wazuh/pull/874))
+- Fix network unreachable error when cluster starts. ([#800](https://github.com/wazuh/wazuh/pull/800))
+
 
 ## [v3.3.1]